Add signal-handling API
As the dbus:publish-objects is a blocking operation (I assume it's reading a
socket to see if any signal was dispatched) I had to make a second thread to run
that operation into. I used stmx to have a thread-safe collection of callbacks.
This thread is only spawned when there's at least a callback defined.